requirements.txt
pyserial~=3.5
SQLAlchemy~=2.0.41
requests~=2.32.3
python-dotenv~=1.1.0